Creada función estatica conect para facilitar hosteos, agregado if con strlen para el unlink de productos, sino daba error en el host

+++ backend/productos.php
@@ -49,10 +49,12 @@ if(isset($_POST['action']) && $_POST['action'] == 'agregar'){/* Con el action so
     if (!$nuevoProducto->guardar()){
         die("En estos momentos, no podemos realizar la operación solicitada");
     } else {
-        try{
-            unlink('../assets/img/productos/'.$_POST['imagen_anterior']);
-        }catch (Exception $ex){
-            echo "<br>la imagen no estaba".$ex;
+        if (strlen($_POST['imagen_anterior']) > 10){
+            try{
+                unlink('../assets/img/productos/'.$_POST['imagen_anterior']);
+            }catch (Exception $ex){
+                echo "<br>la imagen no estaba".$ex;
+            }
         }
         header("location: ".$_SERVER['SCRIPT_NAME']);//esto redirecciona la url al script actual quitando el post
     } 
@@ -65,9 +67,8 @@ if(isset($_POST['action']) && $_POST['action'] == 'agregar'){/* Con el action so
      die();
  }else if (isset($_GET['rem'])) {
     $nuevoProd = new productos($_GET['id']);
-    unlink('../assets/img/productos/'.$_GET['imagen']);
     if ($nuevoProd->eliminar()){
-        //Agregar aca para que borre la imagen
+        unlink('../assets/img/productos/'.$_GET['imagen']);
         header("location: ".$_SERVER['SCRIPT_NAME']);// esto direcciona la url al script actual quitando el GET
     } else{
         die("En estos momenos no podemos eliminar el producto");
